+++ src/simulation/utils/indexing.rs
@@ -28,7 +28,7 @@ pub fn index_to_vector(index: u32, radius: u32) -> IVec3 {
     let y = (index / size % size) - radius;
     let z = (index / area) - radius;
 
-    IVec3::new(x as i32, y as i32, z as i32)
+    IVec3::new(x, y, z)
 }
 
 pub fn vector_to_index(vector: IVec3, radius: u32) -> u32 {
